Almost finished.
The project itself is almost finished and very few things need to be added. I believe I can finish this within another day of work. The last addition is to remove the placeholders and to remove/replace the text background with something else, as well as change the borders. After this, I should be done with the flexbox part of this project.